New monster: bloated husk
Intended for use as an early Swamp enemy, bloated husks charge at
the player at 14 speed and deliver a weakish melee attack. When
killed, they explode, dealing 8d8 physical damage to everything in
radius 2. They're a bit similar to lurking horrors (but much simpler
and easier to deal with) or to ballistomycetes (but with actual HP
and without the very complex spore mechanics).

It might be interesting to have their explosion flatten trees, but
maybe that's just complexity for the sake of it.