Simplify default shift-tab behavior
The previous behavior was to use the quiver only when not in melee range
(and not wielding a launcher). This new behavior seems more appropriate
for the wider range of quiver options now available, including actions
that are untargeted: it simply always triggers the action on shift-tab.
The previous behavior is still present and bindable.  It might be nice
to have better in-game customization of this because I'm not sure
there's one default that really does make sense across all action types.

This is bolted on in a somewhat hacky way to existing autotarget code. A
better refactor might be to extract only the targeting code from
autofight, and have actions use that directly to find a target only when
needed.