Improve autotargeting for various quivered spells
This is the direction chooser's smart autotargeting for most spells on
top of (or in some cases instead of) autofight target selection. There
are a few cases where it is necessary to drop to manual targeting, but
for the most part this allows sensible shift-tab behavior for various
complicated spells like LRD, cloud spells, etc. The simple_targeting
option applies if set.

There's still a somewhat involved division of labor between autofight
target selection and the direction chooser, where afaict the main
difference is that the former prioritizes closer targets, and the latter
prioritizes previous targets if there any, and then closest targets.
Possibly this could all be simplified into one thing? But for now
autofight handles messaging in a useful way for all sorts of error
cases.