Add static targeters for most untargeted spells
Targeters serve two roles: aiming, and telling the player what will
happen once they trigger the action. This is especially brought out with
the revamped fire interface, which shows a targeter for any action, with
a default empty looking-around targeter as a fallback -- this UI works a
lot better if it's actually showing what to do. For many spells this is
exceedingly boring, but for some of the more exotic positional spells I
think it's actually a huge improvement for new players to be able to see
these. In some cases these targeters provide information that was
previously hard to calculate despite being deterministic, with two
standouts being the absolute zero target, and the animate skeleton
target. (If this is bad, then these spells should be randomized when
there are multiple enemies/corpses at the same distance, not
deterministic.)

By default static targeters are not shown for `z`, but there's an option
that lets you force them even there.  No doubt they'd be kind of
annoying to long-time players but this is pretty useful information for
someone who doesn't have every spell memorized.  They are always used
for `Z` casting, and in the fire interface.

Still some cases missing.